requirements.txt
streamlit
pdfplumber
pandas
google-generativeai
xlsxwriter
openpyxl